+++ src/CursorDirection.ts
@@ -1,0 +1,7 @@
+/**
+ * Direction for which way the cursor is moving.
+ */
+export enum CursorDirection {
+  Above = 'above',
+  Below = 'below',
+}