+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## next
+
+- Added the ability to add cursors to the end of each line in a selection.
+
 ## 0.0.1
 
 - First initial release.